Bitvectors: update vmemmap_l_wf

+++ defs.h
@@ -144,9 +144,9 @@ function (boolean) vmemmap_wf (u64 page_index,
     && (page_group_ok(page_index, vmemmap, pool))
 }
 
-function (boolean) vmemmap_l_wf (integer page_index, integer physvirt_offset,
+function (boolean) vmemmap_l_wf (u64 page_index, u64 physvirt_offset,
         pointer virt_ptr,
-        map <integer, struct hyp_page> vmemmap, map <integer, struct list_head> APs,
+        map <u64, struct hyp_page> vmemmap, map <u64, struct list_head> APs,
         pointer pool_pointer, struct hyp_pool pool, excludes ex)
 {
   let page = vmemmap[page_index];
@@ -155,7 +155,7 @@ function (boolean) vmemmap_l_wf (integer page_index, integer physvirt_offset,
   let pool_free_area_pointer = array_shift<struct list_head>(pool_free_area_arr_pointer, page.order);
   let prev = (APs[page_index]).prev;
   let next = (APs[page_index]).next;
-  let free_area_entry = pool.free_area[page.order];
+  let free_area_entry = pool.free_area[(u64) page.order];
   let prev_page_pointer = prev;
   let prev_page_index = cn_hyp_virt_to_pfn(physvirt_offset, prev_page_pointer);
   let prev_page = vmemmap[prev_page_index];
@@ -169,14 +169,14 @@ function (boolean) vmemmap_l_wf (integer page_index, integer physvirt_offset,
         && (prov == (alloc_id) prev)
         && ((APs[prev_page_index]).next == self_node_pointer)
         && (prev_page.order == page.order)
-        && (prev_page.refcount == 0));
+        && (prev_page.refcount == 0u16));
   let next_clause =
     ((next == pool_free_area_pointer) && (free_area_entry.prev == self_node_pointer))
     || (vmemmap_good_pointer (physvirt_offset, next_page_pointer, vmemmap, pool.range_start, pool.range_end, ex)
         && (prov == (alloc_id) next)
         && ((APs[next_page_index]).prev == self_node_pointer)
         && (next_page.order == page.order)
-        && (next_page.refcount == 0));
+        && (next_page.refcount == 0u16));
   // there is no self-loop case for this node type, as it is cleared unless it is
   // present in the per-order free list - TODO delete?
   let nonempty_clause = (prev != self_node_pointer) && (next != self_node_pointer);
